refactor: relocate extensions to root model package
