+++ smithy-kotlin-codegen/src/main/kotlin/software/amazon/smithy/kotlin/codegen/lang/BuiltinPreprocessor.kt
@@ -8,7 +8,7 @@ package software.amazon.smithy.kotlin.codegen.lang
 import software.amazon.smithy.kotlin.codegen.KotlinSettings
 import software.amazon.smithy.kotlin.codegen.core.clientName
 import software.amazon.smithy.kotlin.codegen.integration.KotlinIntegration
-import software.amazon.smithy.kotlin.codegen.model.ext.isEnum
+import software.amazon.smithy.kotlin.codegen.model.isEnum
 import software.amazon.smithy.model.Model
 import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.transform.ModelTransformer
