+++ smithy-kotlin-codegen/src/test/kotlin/software/amazon/smithy/kotlin/codegen/model/OperationNormalizerTest.kt
@@ -8,8 +8,6 @@ package software.amazon.smithy.kotlin.codegen.model
 import io.kotest.matchers.string.shouldContain
 import org.junit.jupiter.api.Test
 import software.amazon.smithy.codegen.core.CodegenException
-import software.amazon.smithy.kotlin.codegen.model.ext.expectShape
-import software.amazon.smithy.kotlin.codegen.model.ext.expectTrait
 import software.amazon.smithy.kotlin.codegen.model.traits.OperationInput
 import software.amazon.smithy.kotlin.codegen.model.traits.OperationOutput
 import software.amazon.smithy.kotlin.codegen.model.traits.SyntheticClone
